bikecompetition/requirements.txt
Django
psycopg2
uwsgi
django-tastypie
django-tastypie-swagger
